temare: [TEST] choose xen subject for xen test
The tests for xen precondition preparation were called without a
given subject. This means that temare selects a topic which was
always xen before because we the database is in a defined state
on test start. Now with the test database changes for autoinstall
xen is no longer the first choice and we have to tell temare to
pick xen when we want to test xen precondition generation.