temare: [FEAT] new subject default to disabled
New subject were inserted as enabled before. This prevents the user to
add completions for this new subject. This in turn may lead to an
incomplete subject being scheduled and unnecessary errors. Now the
default is disabled and whe inform the user that his new subject needs
to be enabled. The better way would be to use a real getopt and add a
switch for enabled/disabled.