Temare: Add and constrain input checks
Add input checks for GRUB template completion keys and values,
and use them in the dbops module.
Move the GRUB template value definition from the config module
to the checks module, because it would create a circular
dependency now, resulting in an import error.
Constrain the subject naming a little bit more, because we
depend on different sub-strings to determine the environment.

Signed-off-by: Frank Arnold <frank.arnold@amd.com>